Правки

+++ app/models/chat.py
@@ -1,4 +1,11 @@
-from sqlalchemy import Column, Integer, String, DateTime
+# -*- coding: utf-8 -*-
+########################
+#    Sokolov Dmitry    #
+# xx.sokolov@gmail.com #
+#  https://t.me/ZbxNTg #
+########################
+# https://github.com/xxsokolov/znt
+from sqlalchemy import Column, Integer, String, DateTime, Numeric
 from sqlalchemy.sql import func
 
 from app.databases.database import Base
@@ -10,8 +17,8 @@ class Chat(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, unique=True, index=True, nullable=False)
-    chat_id = Column(Integer, index=True, nullable=False)
-    chat_id_prev = Column(Integer, index=True)
+    chat_id = Column(Numeric, index=True, nullable=False)
+    chat_id_prev = Column(Numeric, index=True)
     type = Column(String, index=True, nullable=False)
     description = Column(String, index=True)
     time_created = Column(DateTime(timezone=True), server_default=func.now())
